Desktop: Revert link betweeen Employee and Shifts
Now makes deleting a shift dependent on employee again.

Fixes #17
