+++ desktop/src/ca/mcgill/ecse321/foodtruck/model/Shift.java
@@ -5,7 +5,7 @@ package ca.mcgill.ecse321.foodtruck.model;
 import java.sql.Time;
 
 // line 19 "../../../../../FoodTruck.ump"
-// line 57 "../../../../../FoodTruck.ump"
+// line 55 "../../../../../FoodTruck.ump"
 public class Shift
 {
 
@@ -19,23 +19,16 @@ public class Shift
   private Time endTime;
   private double numberOfHours;
 
-  //Shift Associations
-  private Employee employee;
-
   //------------------------
   // CONSTRUCTOR
   //------------------------
 
-  public Shift(String aDayOfWeek, Time aStartTime, Time aEndTime, double aNumberOfHours, Employee aEmployee)
+  public Shift(String aDayOfWeek, Time aStartTime, Time aEndTime, double aNumberOfHours)
   {
     dayOfWeek = aDayOfWeek;
     startTime = aStartTime;
     endTime = aEndTime;
     numberOfHours = aNumberOfHours;
-    if (!setEmployee(aEmployee))
-    {
-      throw new RuntimeException("Unable to create Shift due to aEmployee");
-    }
   }
 
   //------------------------
@@ -94,26 +87,8 @@ public class Shift
     return numberOfHours;
   }
 
-  public Employee getEmployee()
-  {
-    return employee;
-  }
-
-  public boolean setEmployee(Employee aNewEmployee)
-  {
-    boolean wasSet = false;
-    if (aNewEmployee != null)
-    {
-      employee = aNewEmployee;
-      wasSet = true;
-    }
-    return wasSet;
-  }
-
   public void delete()
-  {
-    employee = null;
-  }
+  {}
 
 
   public String toString()
@@ -123,8 +98,7 @@ public class Shift
             "dayOfWeek" + ":" + getDayOfWeek()+ "," +
             "numberOfHours" + ":" + getNumberOfHours()+ "]" + System.getProperties().getProperty("line.separator") +
             "  " + "startTime" + "=" + (getStartTime() != null ? !getStartTime().equals(this)  ? getStartTime().toString().replaceAll("  ","    ") : "this" : "null") + System.getProperties().getProperty("line.separator") +
-            "  " + "endTime" + "=" + (getEndTime() != null ? !getEndTime().equals(this)  ? getEndTime().toString().replaceAll("  ","    ") : "this" : "null") + System.getProperties().getProperty("line.separator") +
-            "  " + "employee = "+(getEmployee()!=null?Integer.toHexString(System.identityHashCode(getEmployee())):"null")
+            "  " + "endTime" + "=" + (getEndTime() != null ? !getEndTime().equals(this)  ? getEndTime().toString().replaceAll("  ","    ") : "this" : "null")
      + outputString;
   }
 }
