+++ desktop/src/ca/mcgill/ecse321/foodtruck/model/Supply.java
@@ -3,8 +3,8 @@
 
 package ca.mcgill.ecse321.foodtruck.model;
 
-// line 33 "../../../../../FoodTruck.ump"
-// line 67 "../../../../../FoodTruck.ump"
+// line 32 "../../../../../FoodTruck.ump"
+// line 65 "../../../../../FoodTruck.ump"
 public class Supply
 {
